Update on "Add Generator(Device device) in C++ API"
This PR adds c-tor `Generator(Device device)` to C++ API to match Python API, to avoid code duplication the logic was moved from `THPGenerator_pynew` to corresponding `at::Generator` c-tor. Unfortunately it makes Generator.h depend on CPUGenerator.h and CUDAGenerator.h, but this is how our Python API works :(

* #34468 Replace Generator* with Generator that holds std::shared_ptr<GeneratorImpl>

[ghstack-poisoned]
